change f64 to f32

+++ Algorithms/src/lib/calculations.rs
@@ -1,5 +1,5 @@
 //assuming three dimensions
-pub fn vector_mul_b(inputs: Vec<f64>, weights: Vec<f64>, bias: f64) -> f64 {
+pub fn vector_mul_b(inputs: Vec<f32>, weights: Vec<f32>, bias: f32) -> f32 {
     let n = inputs.len();
     assert_eq!(inputs.len(), weights.len());
     let mut result = 0.;
@@ -8,7 +8,7 @@ pub fn vector_mul_b(inputs: Vec<f64>, weights: Vec<f64>, bias: f64) -> f64 {
     }
     result + bias
 }
-pub fn _normalize(input: &mut Vec<f64>, data: Vec<f64>) {
+pub fn _normalize(input: &mut Vec<f32>, data: Vec<f32>) {
     assert_eq!(4 * input.len(), data.len());
     for i in 0..input.len() {
         input[i] = (input[i] - data[i * 4]) / (data[i * 4 + 1] + 1e-6).sqrt() * data[i * 4 + 2]
